backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openpyxl>=3.0.0
openpyxl-image-loader>=0.2.1
pydantic>=2.0.0
python-multipart>=0.0.6
pillow>=10.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
